refactor!: move STM etc to ui
