+++ expertsystem/ui/_system_control.py
@@ -10,12 +10,7 @@ from itertools import (
     permutations,
     product,
 )
-from multiprocessing import Pool
-from os import path
 
-from progress.bar import IncrementalBar
-
-import expertsystem
 from expertsystem.state import particle
 from expertsystem.state.particle import (
     CompareGraphElementPropertiesFunctor,
@@ -24,29 +19,17 @@ from expertsystem.state.particle import (
     StateQuantumNumberNames,
     get_interaction_property,
     get_particle_property,
-    initialize_graph,
 )
 from expertsystem.state.propagation import (
-    FullPropagator,
     InteractionNodeSettings,
     InteractionTypes,
 )
 from expertsystem.topology.graph import (
-    InteractionNode,
     StateTransitionGraph,
     get_edges_ingoing_to_node,
-    get_edges_outgoing_to_node,
     get_final_state_edges,
     get_initial_state_edges,
 )
-from expertsystem.topology.topology_builder import (
-    SimpleStateTransitionTopologyBuilder,
-)
-
-from ._default_settings import create_default_interaction_settings
-
-
-_EXPERT_SYSTEM_PATH = path.dirname(path.realpath(expertsystem.__file__))
 
 
 def _change_qn_domain(interaction_settings, qn_name, new_domain):
@@ -68,7 +51,7 @@ def _remove_conservation_law(interaction_settings, cons_law):
             break
 
 
-def _filter_interaction_types(
+def filter_interaction_types(
     valid_determined_interaction_types, allowed_interaction_types
 ):
     int_type_intersection = list(
@@ -95,7 +78,7 @@ class _InteractionDeterminationFunctorInterface(ABC):
         pass
 
 
-class _GammaCheck(_InteractionDeterminationFunctorInterface):
+class GammaCheck(_InteractionDeterminationFunctorInterface):
     """Conservation check for photons."""
 
     name_label = particle.Labels.Name.name
@@ -110,7 +93,7 @@ class _GammaCheck(_InteractionDeterminationFunctorInterface):
         return int_types
 
 
-class _LeptonCheck(_InteractionDeterminationFunctorInterface):
+class LeptonCheck(_InteractionDeterminationFunctorInterface):
     """Conservation check lepton numbers."""
 
     lepton_flavor_labels = [
@@ -153,7 +136,7 @@ class _LeptonCheck(_InteractionDeterminationFunctorInterface):
         return node_interaction_types
 
 
-def _remove_duplicate_solutions(
+def remove_duplicate_solutions(
     results, remove_qns_list=None, ignore_qns_list=None
 ):
     if remove_qns_list is None:
@@ -351,7 +334,7 @@ def _find_node_ids_with_ingoing_particle_name(graph, ingoing_particle_name):
     return found_node_ids
 
 
-def _analyse_solution_failure(violated_laws_per_node_and_graph):
+def analyse_solution_failure(violated_laws_per_node_and_graph):
     # try to find rules that are just always violated
     violated_laws = []
     scoresheet = {}
@@ -378,6 +361,18 @@ def _analyse_solution_failure(violated_laws_per_node_and_graph):
     return violated_laws
 
 
+def create_interaction_setting_groups(graph_node_setting_pairs):
+    graph_settings_groups = {}
+    for (instance, node_settings) in graph_node_setting_pairs:
+        setting_combinations = _create_setting_combinations(node_settings)
+        for setting in setting_combinations:
+            strength = _calculate_strength(setting)
+            if strength not in graph_settings_groups:
+                graph_settings_groups[strength] = []
+            graph_settings_groups[strength].append((instance, setting))
+    return graph_settings_groups
+
+
 def _create_setting_combinations(node_settings):
     return [
         dict(zip(node_settings.keys(), x))
@@ -392,7 +387,7 @@ def _calculate_strength(node_interaction_settings):
     return strength
 
 
-def _match_external_edges(graphs):
+def match_external_edges(graphs):
     if not isinstance(graphs, list):
         raise TypeError("graphs argument is not of type list!")
     if not graphs:
@@ -461,7 +456,7 @@ def _create_edge_id_particle_mapping(graph, external_edge_getter_function):
     }
 
 
-def _perform_external_edge_identical_particle_combinatorics(graph):
+def perform_external_edge_identical_particle_combinatorics(graph):
     """Create combinatorics clones of the `.StateTransitionGraph`.
 
     In case of identical particles in the initial or final state. Only
@@ -522,327 +517,3 @@ def _external_edge_identical_particle_combinatorics(
                 temp_new_graphs.append(graph_copy)
         new_graphs = temp_new_graphs
     return new_graphs
-
-
-class StateTransitionManager:  # pylint: disable=too-many-instance-attributes
-    """Main handler for decay topologies."""
-
-    def __init__(  # pylint: disable=too-many-arguments
-        self,
-        initial_state,
-        final_state,
-        allowed_intermediate_particles=None,
-        interaction_type_settings=None,
-        formalism_type="helicity",
-        topology_building="isobar",
-        number_of_threads=4,
-        propagation_mode="fast",
-    ):
-        if allowed_intermediate_particles is None:
-            allowed_intermediate_particles = []
-        if interaction_type_settings is None:
-            interaction_type_settings = {}
-        allowed_formalism_types = [
-            "helicity",
-            "canonical-helicity",
-            "canonical",
-        ]
-        if formalism_type not in allowed_formalism_types:
-            raise NotImplementedError(
-                f"Formalism type {formalism_type} not implemented."
-                f" Use {allowed_formalism_types} instead."
-            )
-        self.number_of_threads = number_of_threads
-        self.propagation_mode = propagation_mode
-        self.initial_state = initial_state
-        self.final_state = final_state
-        self.interaction_type_settings = interaction_type_settings
-        if not self.interaction_type_settings:
-            self.interaction_type_settings = create_default_interaction_settings(
-                formalism_type
-            )
-        self.interaction_determinators = [_LeptonCheck(), _GammaCheck()]
-        self.allowed_intermediate_particles = allowed_intermediate_particles
-        self.final_state_groupings = []
-        self.allowed_interaction_types = [
-            InteractionTypes.Strong,
-            InteractionTypes.EM,
-            InteractionTypes.Weak,
-        ]
-        self.filter_remove_qns = []
-        self.filter_ignore_qns = []
-        if formalism_type == "helicity":
-            self.filter_remove_qns = [
-                InteractionQuantumNumberNames.S,
-                InteractionQuantumNumberNames.L,
-            ]
-        if "helicity" in formalism_type:
-            self.filter_ignore_qns = [
-                InteractionQuantumNumberNames.ParityPrefactor
-            ]
-        int_nodes = []
-        if topology_building == "isobar":
-            if len(initial_state) == 1:
-                int_nodes.append(InteractionNode("TwoBodyDecay", 1, 2))
-        else:
-            int_nodes.append(
-                InteractionNode(
-                    "NBodyScattering", len(initial_state), len(final_state)
-                )
-            )
-            # turn of mass conservation, in case more than one initial state
-            # particle is present
-            if len(initial_state) > 1:
-                self.interaction_type_settings = create_default_interaction_settings(
-                    formalism_type, False
-                )
-        self.topology_builder = SimpleStateTransitionTopologyBuilder(int_nodes)
-
-        load_default_particle_list()
-
-    def set_topology_builder(self, topology_builder):
-        self.topology_builder = topology_builder
-
-    def add_final_state_grouping(self, fs_group):
-        if not isinstance(fs_group, list):
-            raise ValueError(
-                "The final state grouping has to be of type list."
-            )
-        if len(fs_group) > 0:
-            if not isinstance(fs_group[0], list):
-                fs_group = [fs_group]
-            self.final_state_groupings.append(fs_group)
-
-    def set_allowed_interaction_types(self, allowed_interaction_types):
-        # verify order
-        for allowed_types in allowed_interaction_types:
-            if not isinstance(allowed_types, InteractionTypes):
-                raise TypeError(
-                    "allowed interaction types must be of type"
-                    "[InteractionTypes]"
-                )
-            if allowed_types not in self.interaction_type_settings:
-                logging.info(self.interaction_type_settings.keys())
-                raise ValueError(
-                    f"interaction {allowed_types} not found in settings"
-                )
-        self.allowed_interaction_types = allowed_interaction_types
-
-    def prepare_graphs(self):
-        topology_graphs = self.build_topologies()
-        init_graphs = self.create_seed_graphs(topology_graphs)
-        graph_node_setting_pairs = self.determine_node_settings(init_graphs)
-        # create groups of settings ordered by "probability"
-        graph_settings_groups = self.create_interaction_setting_groups(
-            graph_node_setting_pairs
-        )
-        return graph_settings_groups
-
-    def build_topologies(self):
-        all_graphs = self.topology_builder.build_graphs(
-            len(self.initial_state), len(self.final_state)
-        )
-        logging.info(f"number of topology graphs: {len(all_graphs)}")
-        return all_graphs
-
-    def create_seed_graphs(self, topology_graphs):
-        # initialize the graph edges (initial and final state)
-        init_graphs = []
-        for topology_graph in topology_graphs:
-            topology_graph.set_graph_element_properties_comparator(
-                CompareGraphElementPropertiesFunctor()
-            )
-            init_graphs.extend(
-                initialize_graph(
-                    topology_graph,
-                    self.initial_state,
-                    self.final_state,
-                    self.final_state_groupings,
-                )
-            )
-
-        logging.info(f"initialized {len(init_graphs)} graphs!")
-        return init_graphs
-
-    def determine_node_settings(self, graphs):
-        # pylint: disable=too-many-locals
-        graph_node_setting_pairs = []
-        for graph in graphs:
-            final_state_edges = get_final_state_edges(graph)
-            initial_state_edges = get_initial_state_edges(graph)
-            node_settings = {}
-            for node_id in graph.nodes:
-                node_int_types = []
-                out_edge_ids = get_edges_outgoing_to_node(graph, node_id)
-                in_edge_ids = get_edges_outgoing_to_node(graph, node_id)
-                in_edge_props = [
-                    graph.edge_props[edge_id]
-                    for edge_id in [
-                        x for x in in_edge_ids if x in initial_state_edges
-                    ]
-                ]
-                out_edge_props = [
-                    graph.edge_props[edge_id]
-                    for edge_id in [
-                        x for x in out_edge_ids if x in final_state_edges
-                    ]
-                ]
-                node_props = {}
-                if node_id in graph.node_props:
-                    node_props = graph.node_props[node_id]
-                for int_det in self.interaction_determinators:
-                    determined_interactions = int_det.check(
-                        in_edge_props, out_edge_props, node_props
-                    )
-                    if node_int_types:
-                        node_int_types = list(
-                            set(determined_interactions) & set(node_int_types)
-                        )
-                    else:
-                        node_int_types = determined_interactions
-                node_int_types = _filter_interaction_types(
-                    node_int_types, self.allowed_interaction_types
-                )
-                logging.debug(
-                    "using %s interaction order for node: %s",
-                    str(node_int_types),
-                    str(node_id),
-                )
-                node_settings[node_id] = [
-                    deepcopy(self.interaction_type_settings[x])
-                    for x in node_int_types
-                ]
-            graph_node_setting_pairs.append((graph, node_settings))
-        return graph_node_setting_pairs
-
-    @staticmethod
-    def create_interaction_setting_groups(graph_node_setting_pairs):
-        graph_settings_groups = {}
-        for (graph, node_settings) in graph_node_setting_pairs:
-            setting_combinations = _create_setting_combinations(node_settings)
-            for setting in setting_combinations:
-                strength = _calculate_strength(setting)
-                if strength not in graph_settings_groups:
-                    graph_settings_groups[strength] = []
-                graph_settings_groups[strength].append((graph, setting))
-        return graph_settings_groups
-
-    def find_solutions(
-        self, graph_setting_groups
-    ):  # pylint: disable=too-many-locals
-        """Check for solutions for a specific set of interaction settings."""
-        results = {}
-        logging.info(
-            "Number of interaction settings groups being processed: %d",
-            len(graph_setting_groups),
-        )
-        for strength, graph_setting_group in sorted(
-            graph_setting_groups.items(), reverse=True
-        ):
-            logging.info(
-                "processing interaction settings group with "
-                f"strength {strength}",
-            )
-            logging.info(f"{graph_setting_group} entries in this group")
-            logging.info(f"running with {self.number_of_threads} threads...")
-
-            temp_results = []
-            progress_bar = IncrementalBar(
-                "Propagating quantum numbers...", max=len(graph_setting_group)
-            )
-            progress_bar.update()
-            if self.number_of_threads > 1:
-                with Pool(self.number_of_threads) as pool:
-                    for result in pool.imap_unordered(
-                        self.propagate_quantum_numbers, graph_setting_group, 1
-                    ):
-                        temp_results.append(result)
-                        progress_bar.next()
-            else:
-                for graph_setting_pair in graph_setting_group:
-                    temp_results.append(
-                        self.propagate_quantum_numbers(graph_setting_pair)
-                    )
-                    progress_bar.next()
-            progress_bar.finish()
-            logging.info("Finished!")
-            if strength not in results:
-                results[strength] = []
-            results[strength].extend(temp_results)
-
-        for key, value in results.items():
-            logging.info(
-                f"number of solutions for strength ({key}) "
-                f"after qn propagation: {sum([len(x[0]) for x in value])}",
-            )
-
-        # remove duplicate solutions, which only differ in the interaction qn S
-        results = _remove_duplicate_solutions(
-            results, self.filter_remove_qns, self.filter_ignore_qns
-        )
-
-        node_non_satisfied_rules = []
-        solutions = []
-        for result in results.values():
-            for (tempsolutions, non_satisfied_laws) in result:
-                solutions.extend(tempsolutions)
-                node_non_satisfied_rules.append(non_satisfied_laws)
-        logging.info(f"total number of found solutions: {len(solutions)}")
-        violated_laws = []
-        if len(solutions) == 0:
-            violated_laws = _analyse_solution_failure(node_non_satisfied_rules)
-            logging.info(f"violated rules: {violated_laws}")
-
-        # finally perform combinatorics of identical external edges
-        # (initial or final state edges) and prepare graphs for
-        # amplitude generation
-        _match_external_edges(solutions)
-        final_solutions = []
-        for sol in solutions:
-            final_solutions.extend(
-                _perform_external_edge_identical_particle_combinatorics(sol)
-            )
-
-        return (final_solutions, violated_laws)
-
-    def propagate_quantum_numbers(self, state_graph_node_settings_pair):
-        propagator = self.initialize_qn_propagator(
-            state_graph_node_settings_pair[0],
-            state_graph_node_settings_pair[1],
-        )
-        solutions = propagator.find_solutions()
-        return (solutions, propagator.get_non_satisfied_conservation_laws())
-
-    def initialize_qn_propagator(self, state_graph, node_settings):
-        propagator = FullPropagator(state_graph, self.propagation_mode)
-        for node_id, interaction_settings in node_settings.items():
-            propagator.assign_settings_to_node(node_id, interaction_settings)
-        # specify set of particles which are allowed to be intermediate
-        # particles. If list is empty, then all particles in the default
-        # particle list are used
-        propagator.set_allowed_intermediate_particles(
-            self.allowed_intermediate_particles
-        )
-
-        return propagator
-
-
-def load_default_particle_list() -> None:
-    """Load the default particle list that comes with the expertsystem.
-
-    .. warning::
-        This resets all particle definitions and the removes particles that
-        don't exist in the particle list that ships with the `expertsystem`!
-    """
-    particle_list_file = "particle_list.xml"
-    particle_list_path = path.join(_EXPERT_SYSTEM_PATH, particle_list_file)
-    if not path.exists(particle_list_path):
-        raise FileNotFoundError(
-            f"\n  Failed to load {particle_list_file}!"
-            "\n  Please contact the developers: https://github.com/ComPWA"
-        )
-    particle.DATABASE = dict()
-    particle.load_particles(particle_list_path)
-    logging.info(
-        f"Loaded {len(particle.DATABASE)} particles from {particle_list_file}!"
-    )
