+++ tests/amplitude/test_yaml_helicity.py
@@ -7,7 +7,7 @@ import pytest
 import yaml
 
 from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
-from expertsystem.ui._system_control import (
+from expertsystem.ui import (
     InteractionTypes,
     StateTransitionManager,
 )
